Add a basckground jobs count
I oftent wok with a single terminal for vim and command execution.
Sometime I forget I already have a vim running and start a new one. This
is no efficient as I often have to reopen most of the files I was
working on with the first vim. I hope showing the job count will help me
resume my work with the basckgrounded vim instance when I forgot about
it.