Improve prompt formating in .bashrc
First, this commit uses the bash built in PROMPT_COMMAND variable in
order to set a command that is executed before prompting user for a
command.

Prompt method is cleaned and improved. It now detect if the user is
logged in from ssh and dislplay the hostname in red. Root user username
is colored in red. I implement each PS* prompt definition variable to
have something fancier. PS1 prompt now uses the default '\$' charachter
to display a '$' or '#' depending on user (default bash behavior) but
now it is colored.

As shown in the .bashrc comments, the PS3 implementation is tricky and
works mysteriously as shown by this post forum:
https://www.linuxquestions.org/questions/linux-general-1/bash-prompt-can%27t-change-ps3-4175549239/#post5397699