Fix .bashrc PSI not working properly
When using double quoted string, the shell will substitute the color
variable with the value and the resulting string is used as the PS1
variable. During this operation, the '\$' is converted to '$' and is
defined as is in the PS1 instead of the '\$' string that is use to
change between a '$' or a '#' depending on the user (root, !root). The
solution is, of course, to escape the backslash, resulting the '\$'
string being present and interpreted properly in the prompt.